+++ README.md
@@ -1,8 +1,20 @@
 # smartclock
 Raspberry Pi based smartclock
 
+I use Raspberry Pi version 4 and as screen the Waveshare 4.3" LCD DSI display
+
+## Goal:
+I would like to have a widget based smartclock which has multiple functions.
+First of all the **clock** and a **calendar** with events from Google Calendar. I would like to add **weather forecast** widget as well
+and a public transport schedules widget which shows me when is the next subway/tram/bus departing from a given station. 
+
+## Bill of Material:
+* Raspberry Pi 4B
+* Waveshare 4.3" LCD DSI display
+* 32 GB microSD card
+* USB-C 15W Power supply
 
 ## Resources:
-[Google Calendar API](https://developers.google.com/calendar/api/guides/overview)
-[Qt for Python](https://doc.qt.io/qtforpython-6.2/api.html)
-[All C++ QT classes](https://doc.qt.io/qt-6/classes.html)
+* [Google Calendar API](https://developers.google.com/calendar/api/guides/overview)
+* [Qt for Python](https://doc.qt.io/qtforpython-6.2/api.html)
+* [All C++ QT classes](https://doc.qt.io/qt-6/classes.html)
